money donation (con errores)

+++ anual/src/main/java/org/grupo11/Services/Meal.java
@@ -13,10 +13,10 @@ public class Meal {
     private Long id;
 
     private String type;
-    private int expirationDate;
+    private long expirationDate;
     // @OneToOne //las bidireccionalidades son re de trolo mal
     // private MealDonation mealDonation;
-    private int donationDate;
+    private long donationDate;
     @ManyToOne
     private Fridge fridge;
     private String state;
@@ -41,11 +41,11 @@ public class Meal {
         return type;
     }
 
-    public int getExpirationDate() {
+    public long getExpirationDate() {
         return expirationDate;
     }
 
-    public int getDonationDate() {
+    public long getDonationDate() {
         return donationDate;
     }
